+++ NEWS.md
@@ -1,7 +1,8 @@
 # MiscMetabar 0.9.2 (in development)
 
 - Add param `default_fun` in function `merge_samples2()` in order to replace the default function that change the sample data in case of merging. A useful parameter is `default_fun=diff_fct_diff_class`.
-
+- Add param `kruskal_test` to `hill_pq()` function to prevent user to mis-interpret Tuckey HSD result (and letters) if the global effect of the tested factor on Hill diversity is non significant.
+- Add param `vioplot` to hill_pq() function to allow violin plot instead of boxplot.
 
 # MiscMetabar 0.9.1